refactor(mcp): one implementation for UI and MCP — extract shared services
Per the standing architecture rule (bottleOps/rackOps/statsService precedent):
REST routes and MCP tools must run the SAME code so the two surfaces cannot
drift and changes land in one place.

- services/journalOps.js: sanitizeEntry + validatePairingRefs + createEntry
  (incl. the mention-notification policy) + deleteEntry extracted from
  routes/journal.js; the route AND capture_tasting_note now delegate. The
  tasting-note undo deletes through the same audited path. Tool effects are
  now all-or-nothing: entry first, rating second, compensating delete on a
  late rating failure.
- services/insightsService.js: computeHealthCheck / computeGaps /
  computeValueReport / buildCaseJourneys + loadPortfolio moved out of the MCP
  tool files; mcp/tools/insights.js + journey.js are thin adapters (schemas,
  cache, envelopes). A future UI dead-stock/value page imports these directly.
- services/drinkingService.js: readyCandidates / serializeCandidates /
  scoreDishMatches moved out of mcp/tools/drinking.js (same adapter split).
- services/rackOps.js: buildAnnotatedEntries (withMaturity-parity annotation)
  + applyArrangement (one-save atomic slot assignment + rack.arrange audit)
  — used by auto_arrange apply AND its undo; ready for a future REST arrange
  endpoint. Remaining known duplication: the frontend's own arrange engine
  (utils/rackArrange.js twins) — unifying that means ArrangeModal calling
  server preview/apply, proposed as a follow-up PR.
